Error handling: rename the error container component
