+++ src/App.vue
@@ -1,11 +1,11 @@
 <script setup>
 import { RouterView } from "vue-router";
-import ErrorContainer from "./components/shared/ErrorContainer.vue";
+import HttpStatusContainer from "./components/shared/HttpStatusContainer.vue";
 </script>
 
 <template>
   <div class="font-sans antialiased bg-gray-100 min-h-screen">
-    <ErrorContainer />
+    <HttpStatusContainer />
     <RouterView />
   </div>
 </template>
